HADOOP-15191 patch 003
* fix checkstyle
* add "op_bulk_delete" to the counters
* S3A Distcp test to verify that the bulk delete operation was invoked, so verifying that this path works
* ITestS3ABulkOperations tweaked to handle tests which delete a directory path, to behave differenty under S3Guard. The docs for the (internal) bulk API say 'all entries must be files'';  S3Guard gets inconsistent if you delete a path which has a child entry, whereas S3 classic ignores the request (and if HDFS supported this, would fail with an IOE).

For a public/stable API, we'd need to worry about that problem "deletion of a parent entry"; for s3guard a preflight check of every file not being a dir would be the strategy. Not free, and you'd have to decide what to do: ignore vs reject. Here, because distcp is only deleting files, there's no problem.

Change-Id: Ied48cf78fc4600c8184e2da786bb286bd0dd4399