HADOOP-14556 Patch 024
Isolate MarshalledCredentials (which end up Token Identifiers) from any AWS classes, exceptions, interfaces.
The bridging code  is in a new class, MarshalledCredentialBinding. This makes some bits cleaner, some
bits less so. But it should guarantee that there will never be classload problems with hadoop-aws on the CP
and aws-sdk not. Doesn't mean that you can actually use a DT, but it means that some things like the RM won't
crash on renewals.

Change-Id: I57d75b4ae489c86f3ebe0d380829c94bc602ed8b
Also: AWSCredentialProvider list supports a name field; this is set in the new uses. Helps people understand from an unauthed error that yes, there is a token binding, it just failed to init as no token for this URL was found.