HADOOP-15183:  S3AFileSystem does rename/3
Factored the rename check logic out into a RenameHelper which is then used in S3A FileSystem as  the PoC of how it can use the RenameHelper then directly invoke the inner operations. Added some more @Retry attributes as appropriate.

Conclusion: it works, but for efficient IOPS then `innerRename()` needs to take optional source and dest status values and so so omit repeating the checks.

For more work on this

* the tests; file context has some so review and add to AbstractContractRenameTest. Also, based on some feedback from Sean Mackrory: verify the renamed files actually have the data.
* move internal use of rename (distcp, maprv2 FileOutputcommitter), etc to use this.

of the 50+ places which call rename, they seem split 3-ways int

1. subclasses and forwarding
2. invocations whch check the return value and throw an IOE
3. invocations which are written on the assumption that renames raise exceptions on failure

2 & 3 are the ones to change.

Change-Id: Id77ed7352b9d5ddb124f9191c5c5f1b8a76da7bb